fix: Address Copilot review comments
Addresses all 4 issues from Copilot review:

1. **Refactor duplicate get_current_branch** (copilot_review.rs:269)
   - Extracted to public `git::get_current_branch()` function
   - Removed duplicates from file_pr.rs and copilot_review.rs
   - Both modules now use shared implementation

2. **Fix PR number truncation** (file_pr.rs:201)
   - Changed AgentEvent::PrFiled.pr_number from u32 to u64
   - Prevents data loss for very large PR numbers
   - Matches GitHub's u64-sized IDs

3. **Reduce log noise** (zellij_events.rs:33)
   - Removed internal warning log from emit_event
   - Callers still log warnings, avoiding duplicate log lines
   - Keeps logs cleaner and easier to scan

4. **Fix typo in branch name** (INITIAL_CONTEXT.md:3)
   - Changed "zellij-pip" to "zellij-pipe"
   - Matches actual branch naming convention

All changes compile cleanly with no warnings.

+++ rust/exomonad-ui-protocol/src/lib.rs
@@ -25,7 +25,7 @@ pub enum AgentEvent {
     #[serde(rename = "pr:filed")]
     PrFiled {
         agent_id: String,
-        pr_number: u32,
+        pr_number: u64,
         timestamp: String,
     },
     #[serde(rename = "copilot:reviewed")]
